requirements.txt
paddlepaddle
paddleocr
opencv-python
pillow
numpy
shapely
pyclipper
